+++ packages/core/src/request-builder/response-transformers.ts
@@ -6,6 +6,10 @@ import { Constructable } from '../constructable';
 import { Entity } from '../entity';
 import { deserializeEntity } from '../entity-deserializer';
 
+export async function transformReturnValueForPromiseVoid<ReturnT>(data: any, builderFn: (data: any) => ReturnT) {
+  return builderFn(data);
+}
+
 export function transformReturnValueForEntity<ReturnT extends Entity>(data: any, entityConstructor: Constructable<ReturnT>): ReturnT {
   return deserializeEntity(data.d, entityConstructor).setOrInitializeRemoteState() as ReturnT;
 }