+++ CHANGELOG.md
@@ -29,6 +29,9 @@ blog:
 
 ## fixedIssues
 
+- Fix the generator crashing for services containing function modules without a return type.
+Now the generator will handle these cases and use  `Promise<void>` as return type.
+
 -
 
 # 1.18.0